requirements.txt
boto3 #minio
elasticsearch~=8.19.0
dotenv
kafka-python
pdf2image
pdfplumber
PyMuPDF